[rust] Refactor lock logic in a separate module
